Fixed type concrete strategies to resist their own type. Fixed to return a float not an int.

+++ Assets/__Game/Scripts/CoreComponents/DamageCalcuators(Interfaces)/IAttackTypeDamageCalculation.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public interface IAttackTypeDamageCalculation
 {
-    int CalculateDamageModifier(int damageAmount, AttackType attackerType);
+    float CalculateDamageModifier(float damageAmount, AttackType attackerType);
 }
